+++ packages/terra-framework-docs/src/terra-dev-site/doc/data-grid/FlowsheetDataGrid.4/Examples.6/FlowsheetWithSections.jsx
@@ -4,7 +4,7 @@ import { FlowsheetDataGrid } from 'terra-data-grid';
 const gridDataJSON = {
   cols: [
     {
-      id: 'Column-0', displayName: 'Patient', sortIndicator: 'ascending', isSelectable: true,
+      id: 'Column-0', displayName: 'Patient', isSelectable: true,
     },
     {
       id: 'Column-1', displayName: 'Location', isSelectable: true,