+++ packages/terra-table/src/index.js
@@ -1,5 +1,12 @@
 import Table from './Table';
 import GridContext, { GridConstants } from './utils/GridContext';
+import cellShape from './proptypes/cellShape';
+import columnShape from './proptypes/columnShape';
+import rowShape from './proptypes/rowShape';
+import sectionShape from './proptypes/sectionShape';
+import validateRowHeaderIndex from './proptypes/validators';
 
 export default Table;
-export { GridContext, GridConstants };
+export {
+  GridContext, GridConstants, cellShape, columnShape, rowShape, sectionShape, validateRowHeaderIndex,
+};