+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added
   * Added `hasVisibleColumnHeaders` example for FlowsheetDataGrid.
+  * Added examples and tests to `terra-data-grid` for sections in FlowsheetDataGrid.
 
 * Updated
   * Updated About page of `terra-table` to provide accessibility documentation for Home and End keys.