+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,12 +3,14 @@
 ## Unreleased
 
 * Added
+  * Added new test for Terra Slider component for long field labels.
   * Added `hasVisibleColumnHeaders` example for FlowsheetDataGrid.
-
+  * Added examples and tests to `terra-data-grid` for sections in FlowsheetDataGrid.
+  
 * Updated
   * Updated About page of `terra-table` to provide accessibility documentation for Home and End keys.
   * Removed alpha notice for FlowsheetDataGrid.
-
+  * Removed unsupported props `isMasked` and `isSelectable` from FlowsheetDataGrid subcomponent documentation.
 
 ## 1.44.0 - (November 9, 2023)